refactor(session): extract shared session identifier
